pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "api-drift-agent"
version = "0.1.0"
description = "CLI agent for detecting OpenAPI and FastAPI contract drift."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "coverage>=7,<8",
  "gitpython>=3,<4",
  "jsonschema>=4,<5",
  "libcst>=1.3,<2",
  "ollama>=0.2,<1",
  "PyYAML>=6,<7",
  "rich>=13,<14",
  "ruamel.yaml>=0.18,<1",
  "typer>=0.12,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[project.scripts]
drift-check = "drift_agent.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
markers = [
  "integration: requires local Ollama model/runtime",
]

